# src/Inventory.Api/AutoGen/swagger.json
{
  "x-generator": "NSwag v13.14.8.0 (NJsonSchema v10.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "swagger": "2.0",
  "info": {
    "title": "Inventory",
    "version": "1.0.0"
  },
  "paths": {
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "GetAllProducts",
        "parameters": [
          {
            "type": "string",
            "name": "platform",
            "in": "query",
            "maxLength": 200,
            "x-nullable": true
          },
          {
            "type": "string",
            "name": "productType",
            "in": "query",
            "maxLength": 200,
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "AddProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddProductRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "GetProductById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "operationId": "UpdateProduct",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddProductRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/Stock/{id}": {
      "get": {
        "tags": [
          "Stock"
        ],
        "operationId": "GetStockById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "Stock"
        ],
        "operationId": "UpdateStock",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "format": "guid",
            "x-nullable": false
          },
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateStockRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "404": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/api/Stock": {
      "post": {
        "tags": [
          "Stock"
        ],
        "operationId": "AddStock",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddStockRequest"
            },
            "x-nullable": false
          }
        ],
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    }
  },
  "definitions": {
    "ProblemDetails": {
      "type": "object",
      "additionalProperties": {},
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      }
    },
    "AddProductRequest": {
      "type": "object",
      "required": [
        "name",
        "description",
        "platform",
        "productType"
      ],
      "properties": {
        "name": {
          "type": "string",
          "maxLength": 200,
          "minLength": 1
        },
        "description": {
          "type": "string",
          "maxLength": 255,
          "minLength": 1
        },
        "platform": {
          "type": "string",
          "maxLength": 200,
          "minLength": 1
        },
        "productType": {
          "type": "string",
          "maxLength": 200,
          "minLength": 1
        },
        "releaseDate": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "AddStockRequest": {
      "type": "object",
      "required": [
        "productId",
        "condition",
        "quantity"
      ],
      "properties": {
        "productId": {
          "type": "string",
          "format": "guid"
        },
        "condition": {
          "type": "string",
          "maxLength": 200,
          "minLength": 1
        },
        "quantity": {
          "type": "integer",
          "format": "int32",
          "maximum": 2147483647.0,
          "minimum": 0.0
        }
      }
    },
    "UpdateStockRequest": {
      "type": "object",
      "required": [
        "quantity"
      ],
      "properties": {
        "quantity": {
          "type": "integer",
          "format": "int32",
          "maximum": 2147483647.0,
          "minimum": 0.0
        }
      }
    }
  }
}